src/test/javascript/spec/app/entities/traza-peticion/traza-peticion.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { TrazaPeticionService } from 'app/entities/traza-peticion/traza-peticion.service';
import { ITrazaPeticion, TrazaPeticion, TipoNotificacion, NombreOperacion } from 'app/shared/model/traza-peticion.model';

describe('Service Tests', () => {
    describe('TrazaPeticion Service', () => {
        let injector: TestBed;
        let service: TrazaPeticionService;
        let httpMock: HttpTestingController;
        let elemDefault: ITrazaPeticion;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(TrazaPeticionService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new TrazaPeticion(0, TipoNotificacion.EMAIL, NombreOperacion.ACEPTACION_VERSION_LEY, currentDate, 'AAAAAAA');
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        fecha: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a TrazaPeticion', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        fecha: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        fecha: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new TrazaPeticion(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a TrazaPeticion', async () => {
                const returnedFromService = Object.assign(
                    {
                        tipoNotificacion: 'BBBBBB',
                        nombreOperacion: 'BBBBBB',
                        fecha: currentDate.format(DATE_TIME_FORMAT),
                        observaciones: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        fecha: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of TrazaPeticion', async () => {
                const returnedFromService = Object.assign(
                    {
                        tipoNotificacion: 'BBBBBB',
                        nombreOperacion: 'BBBBBB',
                        fecha: currentDate.format(DATE_TIME_FORMAT),
                        observaciones: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        fecha: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a TrazaPeticion', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
